feat: Complete F64-to-F32 dtype migration across all neural modules
Migrates the entire project's default numerical precision from F64 to F32
while preserving F64 fallback paths where required by numerical stability.

Core infrastructure:
- src/dtype.rs: Add Storage::ptr_eq() for cross-dtype Arc equality
- src/autograd.rs: Default constructors (new, zeros, rand, randn) now
  create F32 tensors. All _generic backward closures converted to
  grad_write_compat() with f64 math to handle mixed-dtype parents.
  Fixed dtype bugs in index_select, transpose, reshape, neg, and
  weighted_mse_loss. Binary save/load uses native dtype bytes.
- src/simd.rs: Remove unused F64 vector_scale paths. F32 inference
  SIMD functions already present.

Neural network modules:
- src/nn.rs: Linear::new creates F32 weights via Tensor::rand_f32.
  RMSNorm::forward backward closure uses grad_write_compat().
- src/transformer.rs: RoPE, batched_matmul_qt_k, batched_matmul_probs_v,
  concat_last_dim, apply_causal_mask, scale_tensor, and softmax all
  converted to dtype-aware paths (no longer hardcode F64). Removed
  unused TensorReadGuard and Dtype imports.
- src/dqn.rs: Scratch buffers, replay cache, and training metrics
  all use F32 accessors. copy_tensor and interpolate made dtype-aware.
- src/ppo.rs: All inference scratch and EMA test code converted to
  data_f32()/data_write_f32().
- src/sim.rs: PPO fast/slow alignment test reads F32 logits.

Application layer:
- src/bench.rs, src/main.rs, data/config.json: Minor F32 adjustments.

Verification:
- cargo test: 130/130 passing
- cargo clippy -- -D warnings: clean
- cargo fmt: applied

+++ src/bench.rs
@@ -376,7 +376,7 @@ fn run_path_comparison(base_config: &Config, seed: u64) -> Vec<(String, Vec<f64>
     let ppo = train_ppo_with_metrics(&mut rng, &env_net, &cfg, None);
 
     let input_dim = crate::neural::DIM;
-    let sample_input: Vec<f64> = (0..input_dim).map(|i| (i as f64) * 0.1 + 0.05).collect();
+    let sample_input: Vec<f32> = (0..input_dim).map(|i| (i as f32) * 0.1 + 0.05).collect();
     let iterations = 2000;
 
     let mut all_latencies: Vec<(String, Vec<f64>)> = Vec::new();
